Rename automated_mcmc_ordering_coupling_copy.py -> mcmc.py
