+++ src/polychron/mcmc.py
